+++ util/iron/Changelog.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Middleware for handler metrics collection.
+- MetricsHandler to expose prometheus metrics.